refactor: move TagCount and CategoryCount types to post-taxonomy module
