+++ src/lib/post-taxonomy.ts
@@ -1,6 +1,17 @@
 // src/lib/post-taxonomy.ts
 import type { PostMeta } from '@/src/types/post';
 
+export type CategoryCount = {
+  category: string;
+  count: number;
+};
+
+export type TagCount = {
+  tag: string;
+  count: number;
+};
+
+
 /**
  * 投稿配列から重複なしのタグ一覧を返す。
  */
@@ -11,7 +22,7 @@ export function getAllTags(posts: PostMeta[]): string[] {
 /**
  * 投稿配列からカテゴリと出現回数の一覧を返す（アルファベット順）。
  */
-export function getCategoriesWithCount(posts: PostMeta[]): { category: string; count: number }[] {
+export function getCategoriesWithCount(posts: PostMeta[]): CategoryCount[] {
   const countMap = new Map<string, number>();
   for (const post of posts) {
     countMap.set(post.category, (countMap.get(post.category) ?? 0) + 1);
@@ -24,7 +35,7 @@ export function getCategoriesWithCount(posts: PostMeta[]): { category: string; c
 /**
  * 投稿配列からタグと出現回数の一覧を返す（出現回数の多い順）。
  */
-export function getTagsWithCount(posts: PostMeta[]): { tag: string; count: number }[] {
+export function getTagsWithCount(posts: PostMeta[]): TagCount[] {
   const countMap = new Map<string, number>();
   for (const post of posts) {
     for (const tag of post.tags) {
